Hecho: * WEBSOCKETS en las tres interfaces. * Turno, cuenta regresiva y fase en la barra lateral.
Cosas que quedan:
* Avisar inconvenientes al usuario. TODO lo que es ERRORES.
* Adornar el FrontEnd

* Hacer responsive la barra superior del front end de gobernadores.

* combates / todo el movimiento militar EFECTIVO / participación de control allí. LABERÍNTICO!!! (Por ahora lo vamos a dejar en tablas como en la versión 26/6/2021)

+++ src/main/java/com/megajuegos/independencia/dao/EjercitosDao.java
@@ -8,11 +8,11 @@ public interface EjercitosDao {
 
     EjercitosModel recursosEjercito(String ciudad);
 
-    void movimientos(String ciudad, EjercitosModel traido, int fase);
+    void movimientos(String ciudad, EjercitosModel traido, long fase);
 
     void asignarUnidades(String ciudad, EjercitosModel traido);
 
-    int fase();
+    long fase();
 
     List<EjercitosModel> listarMovimientos();
